+++ src/sdk/CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Adds Getusereffectivepermissions for IWeb, IList and IListItem #926 [MathijsVerbeeck - Mathijs Verbeeck]
 - Added license information in the generated Nuget packages #945 [clurdish - Landon Petzoldt]
 - `Author` property for `IWeb` #947 [jansenbe - Bert Jansen]
-- Static method to create a PnPContext mock for use in testing #948 [jansenbe - Bert Jansen]
+- `IPnPContext` interface on `PnPContext` to support mocking for use in testing #948 [jansenbe - Bert Jansen]
 - Support for getting, setting and clearing of (folder based) column defaults #949 [jansenbe - Bert Jansen]
 
 ### Changed